Whatapp_Chat_Bot/requirements.txt
# Core Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
flask==3.0.2
pydantic[email]==2.6.1
python-dotenv==1.0.0

# AI and Vector Search
groq==0.18.0
pinecone-client==3.0.1
numpy==1.26.3
tenacity==8.2.3

# Development Dependencies
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
requests-mock==1.11.0